refactor: join SaleButton and RentButton in one component SaleRentButton
